+++ aptihramy-frontend/src/router/index.ts
@@ -3,7 +3,7 @@ import EditPage from '@/pages/EditPage.vue'
 import HomePage from '../pages/HomePage.vue'
 import TrackingChain from '@/pages/TrackingChain.vue'
 import { createRouter, createWebHistory, RouteRecordRaw } from 'vue-router' // Correct import for createRouter
-import EditPageTest from '@/pages/EditPageTest.vue'
+import EditPageTest from '@/pages/EditPage.vue'
 
 const routes: RouteRecordRaw[] = [
   {
